+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+
+- http: Fixed merging user-defined HTTP settings and datasource defaults.
+- tzkt: Fixed iterating over big map keys.
+
 ## [6.4.1] - 2022-12-22
 
 ### Fixed